add deps for styling and base components
