--- .github/workflows/release.yml
@@ -1,30 +1,0 @@
-name: Release
-on:
-  push:
-    branches: 
-      - master
-jobs:
-  release:
-    name: Release
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: zendesk/checkout@v3
-        with:
-          persist-credentials: false
-      - name: Setup Node.js
-        uses: zendesk/setup-node@v3
-        with:
-          node-version-file: '.nvmrc'
-      - name: Install dependencies
-        run: yarn install --frozen-lockfile
-      - name: Build
-        run: yarn build
-      - name: Release
-        run: yarn semantic-release
-        env:
-          GITHUB_TOKEN: ${{ secrets.release_bot_github_token}}
-          GIT_AUTHOR_NAME: ${{ secrets.release_bot_git_name }}
-          GIT_AUTHOR_EMAIL: ${{ secrets.release_bot_git_email }}
-          GIT_COMMITTER_NAME: ${{ secrets.release_bot_git_name }}
-          GIT_COMMITTER_EMAIL: ${{ secrets.release_bot_git_email }}
