+++ .prettierignore
@@ -1,0 +1,5 @@
+.vscode
+.yarn
+dist
+templates
+translations
