+++ manifest.json
@@ -1,5 +1,5 @@
 {
-  "name": "Copenhagen Test",
+  "name": "Copenhagen",
   "author": "Zendesk",
   "version": "4.2.2",
   "api_version": 4,
