+++ src/lib/types.ts
@@ -56,3 +56,10 @@ export type GlobalSettings = {
   };
   connectBlockNewsletterTitle: string;
 };
+
+export interface Navigation {
+  article_id: number;
+  html_url: string;
+  title: string;
+  snippet: string;
+}
