--- src/modules/navigation/data-types/Navigation.ts
@@ -1,6 +1,0 @@
-export interface Navigation {
-  article_id: number;
-  html_url: string;
-  title: string;
-  snippet: string;
-}
