--- src/modules/navigation/data-types/index.ts
@@ -1,1 +1,0 @@
-export * from "./Navigation";
